+++ AppExamples/misc/PytestTemplate/scripts/tests/test_integration/test_blackbox.py
@@ -3,7 +3,7 @@
 This test runs the Unit Under Test (UUT) as script and treats it as opaque. After running it, the ZEISS INSPECT project
 is checked for the expected changes of state (in this example: project keywords).
 
-Carl Zeiss GOM Metrology GmbH, 2025
+Carl Zeiss GOM Metrology GmbH, 2026
 """
 
 import gom