+++ AppExamples/misc/PytestTemplate/scripts/tests/test_integration/test_service.py
@@ -7,14 +7,14 @@ because they run in separate Python interpreter processes.
 Instead, a dedicated coverage setup is used for services -
 see services/reflector.py
 
-Carl Zeiss GOM Metrology GmbH, 2025
+Carl Zeiss GOM Metrology GmbH, 2026
 """
 
 import gom
 import time
 import gom.api.services
 from app_utils.service_manager import ServiceManager
-    
+
 def test_reflect():
     """ Test a service
     
@@ -22,7 +22,7 @@ def test_reflect():
     which is especially useful if the App is not installed, but used
     from a connected folder. 
     """
-    
+
     with ServiceManager("Pytest Reflector"):
         from gom.api.pytest_template.reflect import reflect